Move temporary example code to a separate file.
