+++ HouseRules_Configuration/HouseRules_Configuration.csproj
@@ -55,6 +55,7 @@
     <ItemGroup>
         <Compile Include="ConfigManager.cs" />
         <Compile Include="ConfigurationMod.cs" />
+        <Compile Include="ExampleRulesetWriter.cs" />
         <Compile Include="Properties\AssemblyInfo.cs" />
         <Compile Include="UI\RulesetSelectionUI.cs" />
         <Compile Include="UI\RulesetSelectionPanel.cs" />
